Extract handleEvent
